requirements.txt
pyqt6
python-docx
regex
pandas
openpyxl